Update PyUI.py
added new functions too table, "row_append" which adds a row on the end of the table, and "row_insert" which inserts the row into an inputted part of the table. Each function refreshes only what is necessary making them far more alternative than other solutions 

side note, interesting bug with insert where i have to redo every ID for all the objects below it as they have an id based on the position in the table.