Update PyUI.py
made the default vertical and horizontal spacing variables auto from being defined in the make function to universally in the init of GUI_ITEM, which fixed really blood weird bug where in a table in the make it happen program the title in the edit menu would be randomly split into 2 lines. this is the weirdest buug ever and i hate it.