fixed bug
changed default system to fix bug where when the set default value is x, and the universal or base default is y, if you try and set the value when u make the object to y, it doesnt see any change and sets the value to x. i have chanegd nearly every universal default to -1 or something similar.